fix some bugs and freeze requirements

+++ website/admins/routes.py
@@ -1,31 +1,19 @@
-from __future__ import print_function
-import sys # In python 2.7
-from flask import Blueprint, flash
-from wtforms import PasswordField
 from website import admin, db, bcrypt
+from website.models import User, Category, Course, JoinedCourse
+from website.admins.forms import NewCourseForm, UpdateCourseForm, NewUserForm
+from wtforms import PasswordField
+from flask import Blueprint, flash
 from flask_admin.contrib.sqla import ModelView
 from flask_admin import AdminIndexView
-from website.models import User, Category, Course, JoinedCourse
-from flask_login import current_user
-from website.admins.forms import NewCourseForm, UpdateCourseForm, NewUserForm
 from flask_admin.form.upload import FileUploadField
-
-
-from website.models import Course
-from website.models import User
-
+from flask_login import current_user
 
 
 #for logging use: print('\n\n\n msg', file=sys.stderr)
 
 
-
 ### methods ###
 from website.helper import delete_picture, save_picture
-from website.units.helper import getMaxNumberInUnit
-from website.lessons.helper import choice_query_test, getMaxNumberInLesson 
-
-
 
 
 admins_bp=Blueprint('admins_bp',__name__)
@@ -34,8 +22,6 @@ class MyModelView(ModelView):
   def is_accessible(self):
     #change it to current_user.is_admin but know keep it for testing
     return current_user.is_authenticated and current_user.is_admin == True
-  
-
 
 class MyAdminIndexView(AdminIndexView):
    def is_accessible(self):
@@ -45,14 +31,13 @@ class MyAdminIndexView(AdminIndexView):
    
 ############################## User ################################
    
-class UserAdmin(ModelView):
+class UserAdmin(MyModelView):
   #column_list = ('fname', 'lname', 'email', 'is_instructor', 'is_admin')
   #column_list = ('fname', 'lname', 'email', 'is_instructor')
 
   column_list = ('fname', 'lname', 'email', 'is_instructor', 'is_admin',)
   column_labels = dict(fname = 'First Name', lname = 'Last Name')
   form_excluded_columns = ('is_super_admin','password')
-
 
   form_extra_fields = {
         'file_path': FileUploadField('Profile image',
@@ -61,23 +46,25 @@ class UserAdmin(ModelView):
         'password2': PasswordField('Password')
                       }
   
-  form_columns = (
-        'fname',
-        'lname',
-        'email',
-        'file_path',
-        'password2',
-        'is_instructor',
-        'is_admin',   
-    )
   
-  column_searchable_list = ['fname', 'lname', 'email']
+  form_columns = (
+    'fname',
+    'lname',
+    'email',
+    'file_path',
+    'password2',
+    'is_instructor',
+    'is_admin',   
+    )
+
+  
+  column_searchable_list = ['fname', 'lname']
   page_size = 20
 
   def create_form(self, obj=None):
+    
     return NewUserForm()
   
-
   def delete_model(self, model):
     try:
       if current_user.is_super_admin:
@@ -92,19 +79,21 @@ class UserAdmin(ModelView):
     
     return True
 
-  
   def on_model_change(self, form, model, is_created):
     try:
-      
+
       if form.password2.data != '':
               model.password = bcrypt.generate_password_hash(form.password2.data).decode("utf-8")
+
+
+      '''
       else:
           if not is_created:
             # Retrieve the existing password and retain it
               existing_model = self.session.query(self.model).get(model.id)
               model.password = existing_model.password
-          
-     
+      
+      '''        
       
       if form.email.data != model.email:  # Email changed
        if self.model_class.query.filter_by(email=form.email.data).first():
@@ -112,28 +101,21 @@ class UserAdmin(ModelView):
         return False  # Prevent saving
        
       if not is_created and form.file_path.data:
-       
-        delete_picture(model.img_file, path="static/images/user_pics") 
+
+        if model.img_file != 'default_image.jpg':
+          delete_picture(model.img_file, path="static/images/user_pics") 
 
         filename = save_picture(form.file_path.data, path="static/images/user_pics")
         model.img_file = filename
-  
+
       return super().on_model_change(form, model, is_created)
     
     except Exception as e:
       flash(f"An error occurred while saving the model: {str(e)}", "error")
 
-  
-  def is_accessible(self):
-    #change it to current_user.is_admin but know keep it for testing
-    return current_user.is_authenticated and current_user.is_admin == True
-  
-
-  
-  
 ############################## Course ################################
 
-class CourseAdmin(ModelView):
+class CourseAdmin(MyModelView):
 
   column_list = ('title', 'description', 'price')
   form_excluded_columns = ('icon',)
@@ -195,21 +177,16 @@ class CourseAdmin(ModelView):
                                path="static/images/course_pics")
         model.icon = icon_file
 
-      model.author= current_user
+      model.author= form.instructor.data
 
       return super().on_model_change(form, model, is_created)
     
     except Exception as e:
       flash(f"An error occurred while saving the model: {str(e)}", "error")
 
-  def is_accessible(self):
-    #change it to current_user.is_admin but know keep it for testing
-    return current_user.is_authenticated and current_user.is_admin == True
-  
-
 ############################## Category ################################
 
-class CategoryAdmin(ModelView):
+class CategoryAdmin(MyModelView):
   column_list = ('title',)
   form_excluded_columns = ('icon',)
 
@@ -233,7 +210,7 @@ class CategoryAdmin(ModelView):
   
 ############################# Joined course #################################
 
-class JoinedCourseAdmin(ModelView):
+class JoinedCourseAdmin(MyModelView):
   column_list = ('course.title','user.fname','user.lname' ,'user.email', 'enroll_date','course_progress' )
   column_labels = {"course.title":'Course Title','user.fname':'First Name','user.lname':'Last Name',
                    'user.email':'Email' ,'enroll_date':'Enroll Date'}
@@ -241,12 +218,6 @@ class JoinedCourseAdmin(ModelView):
   form_excluded_columns = ('enroll_date')
 
  
-
-  def is_accessible(self):
-    #change it to current_user.is_admin but know keep it for testing
-    return current_user.is_authenticated and current_user.is_admin == True
-
-
 
 admin.add_view( UserAdmin(User, db.session) )
 admin.add_view( CategoryAdmin(Category, db.session) )
